requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.1
celery == 5.2.7
redis==3.4.1
requests==2.28.1